Move OtherPlayerBoard section
