+++ src/client/components/OtherPlayerBoard/index.ts
@@ -1,0 +1,1 @@
+export * from './OtherPlayerBoard';
